+++ react/features/shared-video2/components/web/SharedVideoDialog.js
@@ -8,7 +8,7 @@ import { translate } from '../../../base/i18n';
 import { getFieldValue } from '../../../base/react';
 import { connect } from '../../../base/redux';
 import { defaultSharedVideoLink } from '../../constants';
-import { getYoutubeLink } from '../../functions';
+import { getPeerTubeLink } from '../../functions';
 import AbstractSharedVideoDialog from '../AbstractSharedVideoDialog';
 
 /**
@@ -48,7 +48,7 @@ class SharedVideoDialog extends AbstractSharedVideoDialog<*> {
 
         this.setState({
             value: linkValue,
-            okDisabled: !getYoutubeLink(linkValue)
+            okDisabled: !getPeerTubeLink(linkValue)
         });
     }
