+++ react/features/shared-video2/constants.js
@@ -4,7 +4,7 @@
  * Example shared video link.
  * @type {string}
  */
-export const defaultSharedVideoLink = 'https://peer.tube/videos/watch/ae3d7bac-e746-45cd-b4a5-a7b314f20a85';
+export const defaultSharedVideoLink2 = 'https://peer.tube/videos/watch/ae3d7bac-e746-45cd-b4a5-a7b314f20a85';
 
 /**
  * Fixed name of the video player fake participant.
@@ -16,4 +16,4 @@ export const VIDEO_PLAYER_PARTICIPANT_NAME2 = 'PeerTube';
  * Shared video command.
  * @type {string}
  */
-export const SHARED_VIDEO = 'shared-video2';
+export const SHARED_VIDEO2 = 'shared-video2';
