+++ react/features/shared-video2/reducer.native.js
@@ -2,16 +2,16 @@
 
 import { ReducerRegistry } from '../base/redux';
 
-import { SET_SHARED_VIDEO_STATUS } from './actionTypes';
+import { SET_SHARED_VIDEO_STATUS2 } from './actionTypes';
 
 /**
  * Reduces the Redux actions of the feature features/shared-video.
  */
-ReducerRegistry.register('features/shared-video', (state = {}, action) => {
+ReducerRegistry.register('features/shared-video2', (state = {}, action) => {
     const { videoId, status, time, ownerId } = action;
 
     switch (action.type) {
-    case SET_SHARED_VIDEO_STATUS:
+    case SET_SHARED_VIDEO_STATUS2:
         return {
             ...state,
             videoId,
