+++ react/features/shared-video2/reducer.web.js
@@ -2,7 +2,7 @@
 
 import { ReducerRegistry } from '../base/redux';
 
-import { SET_SHARED_VIDEO_STATUS, SET_DISABLE_BUTTON } from './actionTypes';
+import { SET_SHARED_VIDEO_STATUS2, SET_DISABLE_BUTTON2 } from './actionTypes';
 
 /**
  * Reduces the Redux actions of the feature features/shared-video.
@@ -11,13 +11,13 @@ ReducerRegistry.register('features/shared-video2', (state = {}, action) => {
     const { status, disabled } = action;
 
     switch (action.type) {
-    case SET_SHARED_VIDEO_STATUS:
+    case SET_SHARED_VIDEO_STATUS2:
         return {
             ...state,
             status
         };
 
-    case SET_DISABLE_BUTTON:
+    case SET_DISABLE_BUTTON2:
         return {
             ...state,
             disabled
